+++ malachite-base/src/test_util/bench/bucketers.rs
@@ -1065,6 +1065,31 @@ pub fn u64_polynomial_bit_bucketer(var_name: &str) -> Bucketer<'_, U64Polynomial
     }
 }
 
+pub fn pair_u64_polynomial_max_bit_bucketer<'a>(
+    x_name: &'a str,
+    y_name: &'a str,
+) -> Bucketer<'a, (U64Polynomial, U64Polynomial)> {
+    Bucketer {
+        bucketing_function: &|(p, q)| {
+            usize::exact_from(max(
+                p.coefficients_asc()
+                    .iter()
+                    .copied()
+                    .map(SignificantBits::significant_bits)
+                    .sum::<u64>(),
+                q.coefficients_asc()
+                    .iter()
+                    .copied()
+                    .map(SignificantBits::significant_bits)
+                    .sum::<u64>(),
+            ))
+        },
+        bucketing_label: format!(
+            "max({x_name}'s total coefficient bits, {y_name}'s total coefficient bits)"
+        ),
+    }
+}
+
 pub fn pair_1_u64_polynomial_bit_bucketer<T>(var_name: &str) -> Bucketer<'_, (U64Polynomial, T)> {
     Bucketer {
         bucketing_function: &|(p, _)| {