+++ malachite-base/tests/lib.rs
@@ -878,6 +878,9 @@ pub mod u64_polynomial {
         pub mod named;
         pub mod size;
     }
+    pub mod comparison {
+        pub mod cmp;
+    }
     pub mod conversion {
         pub mod from_coefficients_asc;
         pub mod from_u64;