+++ malachite-nz/tests/lib.rs
@@ -559,6 +559,10 @@ pub mod integer_polynomial {
         pub mod named;
         pub mod size;
     }
+    pub mod comparison {
+        pub mod cmp;
+        pub mod shortlex_cmp;
+    }
     pub mod conversion {
         pub mod from_coefficients_asc;
         pub mod from_integer;